% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET_doc.R
\docType{data}
\name{binaryY_bivariateM}
\alias{binaryY_bivariateM}
\title{A Sample Data for Analysis}
\format{
## `binaryY_bivariateM`
A data frame with 500 rows and 5 columns:
\describe{
  \item{A}{Binary treatment variable}
  \item{M}{Bivariate Normally distributed mediator variable}
  \item{Y}{Binary outcome variable}
  \item{X}{Uniformly distributed measure confounder within the range of (0,1)}
  \item{U}{Normally distributed unmeasured confounder}
}
}
\usage{
binaryY_bivariateM
}
\description{
A Sample Data for Analysis
}
\examples{
\donttest{
# data generated with the following Data Generating Process (DGP)
generate_data <- function(n,parA = c(0.3,0.2), parU=c(1,1,1,0), parM = matrix(c(1, 1, 1, 0,-1,-0.5,2,0), nrow = 2,byrow = T), parY = c(1, 1, -0.5,1, 0), sd.U=1){

 ########################################################
 # M is bivariate normal with mean parameter be
 #  1  1.0    1    0
 # -1 -0.5    2    0

 # and the variance covariance matrix be
 # 2 1
 # 1 3
 ########################################################

 X <- runif(n, 0, 1) # p(X)

 A <- rbinom(n, 1, (parA[1] + parA[2]*X)) # p(A|X)

 U <- parU[1] + parU[2]*A + parU[3]*X + parU[4]*A*X + rnorm(n,0,sd.U) # p(U|A,X)

 M <- cbind(parM[1,1] + parM[1,2]*A + parM[1,3]*X + parM[1,4]*A*X,
            parM[2,1] + parM[2,2]*A + parM[2,3]*X + parM[2,4]*A*X)+ mvrnorm(n , mu =c(0,0) , Sigma = matrix(c(2, 1, 1, 3), nrow = 2))

 Y <- rbinom(n,1,plogis(parY[1]*U + parY[2]*M[,1]+ parY[3]*M[,2] + parY[4]*X + parY[5]*M[,1]*X)) # p(Y|U,M,X)

 data <- data.frame(X=X, U=U, A=A, M=M, Y=Y)

 # propensity score
 ps <- A*(parA[1] + parA[2]*X)+(1-A)*(1-(parA[1] + parA[2]*X))

 # mediator density ratio: p(M|a,X)/p(M|A,X)
 f.m.ratio.a1 <- function(i){
   dmvnorm(x=M[i,], mean=cbind(parM[1,1] + parM[1,2]*1 + parM[1,3]*X[i] + parM[1,4]*1*X[i],
                               parM[2,1] + parM[2,2]*1 + parM[2,3]*X[i] + parM[2,4]*1*X[i]),sigma=matrix(c(2, 1, 1, 3), nrow = 2))/
     dmvnorm(x=M[i,], mean=cbind(parM[1,1] + parM[1,2]*A[i] + parM[1,3]*X[i] + parM[1,4]*A[i]*X[i], parM[2,1] + parM[2,2]*A[i] + parM[2,3]*X[i] + parM[2,4]*A[i]*X[i]),sigma=matrix(c(2, 1, 1, 3), nrow = 2))
 }
 f.m.ratio.a0 <- function(i){
   dmvnorm(x=M[i,], mean=cbind(parM[1,1] + parM[1,2]*0 + parM[1,3]*X[i] + parM[1,4]*0*X[i],
                               parM[2,1] + parM[2,2]*0 + parM[2,3]*X[i] + parM[2,4]*0*X[i]),sigma=matrix(c(2, 1, 1, 3), nrow = 2))/
     dmvnorm(x=M[i,], mean=cbind(parM[1,1] + parM[1,2]*A[i] + parM[1,3]*X[i] + parM[1,4]*A[i]*X[i], parM[2,1] + parM[2,2]*A[i] + parM[2,3]*X[i] + parM[2,4]*A[i]*X[i]),sigma=matrix(c(2, 1, 1, 3), nrow = 2))
 }

 m.ratio.a1 <- sapply(1:n, f.m.ratio.a1)
 m.ratio.a0 <- sapply(1:n, f.m.ratio.a0)

 return(list(data = data,
             parA=parA,
             parU=parU,
             parM=parM,
             parY=parY,
             sd.U=sd.U,
             sigma.M=matrix(c(2, 1, 1, 3), nrow = 2),
             ps=ps,
             m.ratio.a1=m.ratio.a1,
             m.ratio.a0=m.ratio.a0))
 }

}

}
\keyword{datasets}
