% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET_doc.R
\docType{data}
\name{continuousY_continuousM_10dX}
\alias{continuousY_continuousM_10dX}
\title{A Sample Data for Analysis}
\format{
## `continuousY_continuousM_10dX`
A data frame with 500 rows and 14 columns:
\describe{
  \item{A}{Binary treatment variable}
  \item{M}{Quadrivariate Normally distributed mediator variable}
  \item{Y}{Normally distributed outcome variable}
  \item{X.i}{10 Uniformly distributed measure confounder within the range of (0,1)}
  \item{U}{Normally distributed unmeasured confounder}
}
}
\usage{
continuousY_continuousM_10dX
}
\description{
A Sample Data for Analysis
}
\examples{
\donttest{
# data generated with the following Data Generating Process (DGP)
generate_data <- function(n,parA = c(0.48, 0.07, 1.00, -1.00, -0.34, -0.12, 0.30, -0.35, 1.00, -0.10, 0.46, # linear terms
 0.33, 0.00, 0.45, 0.1, -0.32, -0.08, -0.2, 0.50, 0.50, -0.03)*0.1,# X^2 high order terms

 parU=c(-2.0, -1.0, -1.0, 2.0, 3.0, 0.5, 3.0, 2.0, -1.0, 1.0, -3.0, 1.5, # linear terms
        -3.0, -2.0,  1.0,  3.0,  1.5), # A*X interactions

 parM = c(3.0, 1.5, -1.5, -1.5, -1.0, -2.0, -3.0, -3.0, -1.5, 2.0, 1.5, 3.0, # linear terms
          1.5, 2.0, 0.5, 0.5, 3.0, # A*X interactions
          -0.2, -0.33, 0.5, 0.3, -0.5)*0.025, # X^2 high order terms

 parY = c(1.0, -2.0, -3.0, -1.5, 1.0, 0.5, -2.0, 1.5, -2.0, -3.0, -3.0, -1.5, # linear terms
          -1.0, 0.5, 3.0, 1.5, 0.5, # M*X interactions
          3, # M^2 high order term
          1.0, 1.5, -2.0, 3.0, -1.0), # X^2 high order terms
 sd.M=1, sd.U=1, sd.Y=1){

   # generate X from uniform distr
   X <- replicate(10, runif(n,0,1))

   A <- rbinom(n, 1, plogis(parA[1] + rowSums(sweep(X, 2, parA[2:11], "*")) +  rowSums(sweep(X^2,2,parA[12:21],"*")) )) # p(A|X)

   U <- parU[1] + parU[2]*A + rowSums(sweep(X, 2, parU[3:12], "*")) + rowSums(diag(A)\%*\%sweep(X[,1:5],2,parU[13:17],"*")) + rnorm(n,0,sd.U) # p(U|A,X)

   M <- parM[1] + parM[2]*A + rowSums(sweep(X, 2, parM[3:12], "*")) + rowSums(diag(A)\%*\%sweep(X[,1:5],2,parM[13:17],"*")) + rowSums(sweep(X[,6:10]^2,2,parM[18:22],"*"))  + rnorm(n,0,sd.M) # p(M|A,X)

   Y <- parY[1]*U + parY[2]*M + rowSums(sweep(X, 2, parY[3:12], "*")) + rowSums(diag(M)\%*\%sweep(X[,1:5],2,parY[13:17],"*")) + parY[18]*M^2 + rowSums(sweep(X[,6:10]^2,2,parY[19:23],"*")) + rnorm(n, 0, sd.Y) # p(Y|U,M,X)

   data <- data.frame(X=X, U=U, A=A, M=M, Y=Y)

   # propensity score
   ps <- A*plogis(parA[1] + rowSums(sweep(X, 2, parA[2:11], "*")) +  rowSums(sweep(X^2,2,parA[12:21],"*")) )+(1-A)*(1-plogis(parA[1] + rowSums(sweep(X, 2, parA[2:11], "*")) +  rowSums(sweep(X^2,2,parA[12:21],"*")) ))

   # mediator density ratio: p(M|a,X)/p(M|A,X)
   m.ratio.a1 <- dnorm(M,parM[1] + parM[2]*1 + rowSums(sweep(X, 2, parM[3:12], "*")) + rowSums(diag(rep(1,n))\%*\%sweep(X[,1:5],2,parM[13:17],"*")) + rowSums(sweep(X[,6:10]^2,2,parM[18:22],"*")) ,sd.M)/
     dnorm(M,parM[1] + parM[2]*A + rowSums(sweep(X, 2, parM[3:12], "*")) + rowSums(diag(A)\%*\%sweep(X[,1:5],2,parM[13:17],"*")) + rowSums(sweep(X[,6:10]^2,2,parM[18:22],"*")) ,sd.M)
   m.ratio.a0 <- dnorm(M,parM[1] + parM[2]*0 + rowSums(sweep(X, 2, parM[3:12], "*")) + rowSums(diag(rep(0,n))\%*\%sweep(X[,1:5],2,parM[13:17],"*")) + rowSums(sweep(X[,6:10]^2,2,parM[18:22],"*")) ,sd.M)/
     dnorm(M,parM[1] + parM[2]*A + rowSums(sweep(X, 2, parM[3:12], "*")) + rowSums(diag(A)\%*\%sweep(X[,1:5],2,parM[13:17],"*")) + rowSums(sweep(X[,6:10]^2,2,parM[18:22],"*")) ,sd.M)

   return(list(data = data,
               X=X,
               parA=parA,
               parU=parU,
               parM=parM,
               parY=parY,
               sd.U=sd.U,
               sd.Y=sd.Y,
               sd.M=sd.M,
               ps=ps,
               m.ratio.a1=m.ratio.a1,
               m.ratio.a0=m.ratio.a0
   ))
 }


}

}
\keyword{datasets}
