% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMLE-binary.R
\name{TMLE.binary}
\alias{TMLE.binary}
\title{Function for implementing TMLE for univariate binary mediators.}
\usage{
TMLE.binary(
  a,
  data,
  treatment,
  mediator,
  outcome,
  covariates,
  onestep = T,
  superlearner = T,
  crossfit = F,
  K = 5,
  lib = c("SL.glm", "SL.earth", "SL.ranger", "SL.mean"),
  n.iter = 500,
  cvg.criteria = 0.01,
  formulaY = "Y ~ .",
  formulaA = "A ~ .",
  formulaM = "M~.",
  linkY_binary = "logit",
  linkA = "logit",
  linkM_binary = "logit",
  truncate_lower = 0,
  truncate_upper = 1
)
}
\arguments{
\item{a}{Treatment level at which the average counterfactual outcome is computed}

\item{data}{A Dataframe contains treatment, mediators, outcome, and measured confounders}

\item{treatment}{Variable name for the unvariate binary treatment}

\item{mediator}{Variable name for the binary univariate mediator}

\item{outcome}{Variable name for the continuous univariate outcome}

\item{covariates}{Variable name for the measured confounders}

\item{onestep}{A logical indicator determines whether one-step estimation is executed. When 'onestep=T', the one-step estimation result is provided.
Conversely, if 'onestep=F', the result is withheld.}

\item{superlearner}{A logical indicator determines whether SuperLearner is adopted for estimating the outcome regression, mediator density, and the propensity score.}

\item{crossfit}{A logical indicator determines whether SuperLearner+Cross-fitting is adopted for estimating the outcome regression, mediator density, and the propensity score.}

\item{K}{A integer indicating the number of folds for cross-fitting, the default is 5.}

\item{lib}{Library of algorithms for SuperLearner or SuperLearner+Cross-fitting.}

\item{n.iter}{The maximum number of iterations performed when iteratively updating the mediator density and propensity score.}

\item{cvg.criteria}{A numerical value representing the convergence criteria when iteratively updating the mediator density and propensity score. The default value is 0.01, meaning update stops when \eqn{max(|\Phi_M|,|\Phi_A|)<0.01}.}

\item{formulaY}{Regression formula for the outcome regression of Y on M, A, X. The default is 'Y ~ 1+ M + A + X'.}

\item{formulaA}{Regression formula for the propensity score regression of A on X. The default is 'A ~ 1 + X'.}

\item{formulaM}{Regression formula for the mediator density regression of M on A and X. The default is 'M ~ 1 + A + X'. This parameter is only needed when M is a univariate binary mediator.}

\item{linkY_binary}{The link function used for the logistic regression of Y on M, A, X when Y is binary. The default is the 'logit' link.}

\item{linkA}{The link function used for the logistic regression of A on X. The default is the 'logit' link.}

\item{linkM_binary}{The link function used for the logistic regression of M on A and X. The default is the 'logit' link. This parameter is only needed when M is a univariate binary mediator.}

\item{truncate_lower}{A numeric variable, setting lower bound for the truncated propensity score. The default is 0.}

\item{truncate_upper}{A numeric variable, setting upper bound for the truncated propensity score. The default is 1.}
}
\value{
a list of initialization of matrices.

Function outputs a list containing TMLE output (and Onestep estimator output if 'onestep=T' is specified):
\describe{
      \item{\code{estimated_psi}}{The estimated parameter of interest: \eqn{E(Y^a)}}
      \item{\code{lower.ci}}{Lower bound of the 95\% confidence interval for \code{estimated_psi}}
      \item{\code{upper.ci}}{Upper bound of the 95\% confidence interval for \code{estimated_psi}}
      \item{\code{theta_x}}{\eqn{\int E(Y|M,A,X)p(M|A=a,X)p(A|X) dM dA}}
      \item{\code{p.m1.aX}}{\eqn{\int p(M=1|A=a,X)}}
      \item{\code{p.a1.X}}{\eqn{p(A=1|X)}}
      \item{\code{or_pred}}{\eqn{E(Y|M,A,X)}}
      \item{\code{EIF}}{The estimated efficient influence function evaluated at the observed data}
      \item{\code{EDstar}}{A vector of the mapping of \code{EIF} in the tangent space of \eqn{Y|M,A,X}, \eqn{M|A,X}, and \eqn{A|X}.}
      \item{\code{EDstar_M.vec}}{A vector containing the average value of the mapping of EIF in tangent space \eqn{M|A,X} over iterations. This is useful for checking the convergence behavior of the mediator density. It's expected to be close to 0 when convergence is achieved.}
      \item{\code{EDstar_ps.vec}}{A vector containing the average value of the mapping of EIF in tangent space \eqn{A|X} over iterations. This is useful for checking the convergence behavior of the propensity score. It's expected to be close to 0 when convergence is achieved.}
      \item{\code{eps2_vec}}{A vector containing the index for submodels of the mediator density over iterations. This is useful for checking the convergence behavior of the mediator density. It's expected to be close to 0 when convergence is achieved.}
      \item{\code{eps3_vec}}{A vector containing the index for submodels of the propensity score over iterations. This is useful for checking the convergence behavior of the propensity score. It's expected to be close to 0 when convergence is achieved.}
      \item{\code{iter}}{Number of iterations where convergence is achieved for the iterative update of the mediator density and propensity score.}}
}
\description{
Function for implementing TMLE for univariate binary mediators.
}
\examples{
\donttest{
TMLE.binary(a=1,data,treatment="A", mediators="M", outcome="Y", covariates="X")
}
}
