% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMLE-np-dnorm.R
\name{TMLE.np.dnorm}
\alias{TMLE.np.dnorm}
\title{Function for implementing TMLE with the mediator density \eqn{p(M|A,X)} estimated assuming normal distribution.}
\usage{
TMLE.np.dnorm(
  a,
  data,
  treatment = "A",
  mediator = "M",
  outcome = "Y",
  covariates = c("X1", "X2", "X3"),
  onestep = T,
  n.iter = 500,
  eps = T,
  cvg.criteria = 0.01,
  formulaY = "Y ~ .",
  formulaA = "A ~ .",
  linkA = "logit",
  truncate_lower = 0,
  truncate_upper = 1
)
}
\arguments{
\item{a}{Treatment level at which the average counterfactual outcome is computed}

\item{data}{A Dataframe contains treatment, mediators, outcome, and measured confounders}

\item{treatment}{Variable name for the unvariate binary treatment}

\item{mediator}{Variable name for the continuous univariate mediator}

\item{outcome}{Variable name for the continuous univariate outcome}

\item{covariates}{Variable name for the measured confounders}

\item{onestep}{A logical indicator determines whether one-step estimation is executed. When 'onestep=T', the one-step estimation result is provided. Conversely, if 'onestep=F', the result is withheld.}

\item{n.iter}{The maximum number of iterations performed when iteratively updating the mediator density and propensity score.}

\item{eps}{A logical indicator determines the stopping criteria used when iteratively updating the mediator density and propensity score. The default is 'eps=T'.
When 'eps=T', \eqn{\sqrt{\epsilon_2^2+\epsilon_3^2}} is used. When 'eps=F', \eqn{max(|\Phi_M|,|\Phi_A|} is used. In general, adoption of 'eps=T' results in better convergence while it takes longer time.
Conversely, adoption of 'eps=F' usually requires less time but can result in algorithm divergence.}

\item{cvg.criteria}{A numerical value representing the convergence criteria when iteratively updating the mediator density and propensity score. The default value is 0.01, meaning update stops when stopping criteria < 0.01. The stopping criteria is chosen by the eps.}

\item{formulaY}{Regression formula for the outcome regression of Y on M, A, X. The default is 'Y ~ 1+ M + A + X'.}

\item{formulaA}{Regression formula for the propensity score regression of A on X. The default is 'A ~ 1 + X'.}

\item{linkA}{The link function used for the logistic regression of A on X. The default is the 'logit' link.}

\item{truncate_lower}{A numeric variable, setting lower bound for the truncated propensity score. The default is 0.}

\item{truncate_upper}{A numeric variable, setting upper bound for the truncated propensity score. The default is 1.}
}
\value{
Function outputs a list containing TMLE output (and Onestep estimator output if 'onestep=T' is specified):
\describe{
      \item{\code{estimated_psi}}{The estimated parameter of interest: \eqn{E(Y^a)}}
      \item{\code{lower.ci}}{Lower bound of the 95\% confidence interval for \code{estimated_psi}}
      \item{\code{upper.ci}}{Upper bound of the 95\% confidence interval for \code{estimated_psi}}
      \item{\code{theta_x}}{\eqn{\int E(Y|M,A,X)p(M|A=a,X)p(A|X) dM dA}}
      \item{\code{p.m1.aX}}{\eqn{\int p(M=1|A=a,X)}}
      \item{\code{p.a1.X}}{\eqn{p(A=1|X)}}
      \item{\code{or_pred}}{\eqn{E(Y|M,A,X)}}
      \item{\code{EIF}}{The estimated efficient influence function evaluated at the observed data}
      \item{\code{EDstar}}{A vector of the mapping of \code{EIF} in the tangent space of \eqn{Y|M,A,X}, \eqn{M|A,X}, and \eqn{A|X}.}
      \item{\code{EDstar_M.vec}}{A vector containing the average value of the mapping of EIF in tangent space \eqn{M|A,X} over iterations. This is useful for checking the convergence behavior of the mediator density. It's expected to be close to 0 when convergence is achieved.}
      \item{\code{EDstar_ps.vec}}{A vector containing the average value of the mapping of EIF in tangent space \eqn{A|X} over iterations. This is useful for checking the convergence behavior of the propensity score. It's expected to be close to 0 when convergence is achieved.}
      \item{\code{eps2_vec}}{A vector containing the index for submodels of the mediator density over iterations. This is useful for checking the convergence behavior of the mediator density. It's expected to be close to 0 when convergence is achieved.}
      \item{\code{eps3_vec}}{A vector containing the index for submodels of the propensity score over iterations. This is useful for checking the convergence behavior of the propensity score. It's expected to be close to 0 when convergence is achieved.}
      \item{\code{iter}}{Number of iterations where convergence is achieved for the iterative update of the mediator density and propensity score.}}
}
\description{
Function for implementing TMLE with the mediator density \eqn{p(M|A,X)} estimated assuming normal distribution.
}
\examples{
\donttest{
TMLE.np(1,data,treatment="A", mediators="M", outcome="Y", covariates=c("X1","X2"))
}
}
